Improve alfasim .xlsl trend value reader
