+++ tests/fixtures/datasets.py
@@ -2,8 +2,9 @@ import os
 import random
 from pathlib import Path
 
-import pandas as pd
 import pytest
+
+from src.xls_reader.xls_reader import merge_tabs
 
 
 def shuffle_dataframe(dataframe):
@@ -43,23 +44,6 @@ def shuffle_dataframe(dataframe):
         dataframe.iloc[:, i] = dataframe_copy.iloc[:, shuffled_order_global[i]]
 
     dataframe.columns = shuffled_keys
-    return dataframe
-
-
-def merge_tabs(xls_file):
-    xls = pd.ExcelFile(xls_file)
-    dataframe = pd.DataFrame()
-
-    for sheet_name in xls.sheet_names:
-        df = xls.parse(sheet_name, decimal=",")
-        dataframe = pd.concat(
-            [dataframe, df],
-            axis=1,
-        )
-    dataframe.to_excel("temporary_file.xlsx", index=False)
-    excel_file = "temporary_file.xlsx"  # TODO: Fix it!
-    dataframe = pd.read_excel(excel_file)
-    os.remove("temporary_file.xlsx")  # TODO: Fix it!
     return dataframe
